Use long for vfs ids instead of uints.

+++ vfs/vfs.core/IJCDBasicVFS.cs
@@ -149,7 +149,7 @@ namespace vfs.core
 
         JCDDirEntry GetFileDetails(string vfsPath);
 
-        int GetId();
-        void SetId(int id);
+        long GetId();
+        void SetId(long id);
     }
 }
